requirements.txt
amqp==5.2.0
asgiref==3.7.2
asttokens==2.4.1
attrs==23.1.0
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
Django==4.2.7
django-celery-results==2.5.1
django-dotenv==1.4.2
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.26.5
executing==2.0.1
inflection==0.5.1
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kombu==5.3.4
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.6
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
redis==5.0.1
referencing==0.31.1
rpds-py==0.13.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tornado==6.4
traitlets==5.14.0
tzdata==2023.3
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.12
